3.落点范围 （完） 4.小人和金币的碰撞（） 5.敌人AI攻击间隔时间和AI的角色创建间隔  速度增加间隔  速度增加增量  血量增加间隔 血量增加增量 6.AI本地规避优化

+++ Assets/Scripts/Bean/MVC/Game/ShipDataBean.cs
@@ -9,6 +9,6 @@ public class ShipDataBean : BaseBean
     public int ship_damage;
 
     public int bulletDamage = 1;
-    public int intervalForFire = 10;
+    public float intervalForFire = 10;
     public CharacterTypeEnum characterType;
 }
